src/app/components/login/login.component.ts
import { AuthService } from './../../services/auth.service';
import { CommonModule } from '@angular/common';
import { FormsModule } from '@angular/forms';
import { Component } from '@angular/core';
import { Router } from '@angular/router';

@Component({
  selector: 'app-login',
  standalone: true,
  imports: [CommonModule, FormsModule],
  templateUrl: './login.component.html',
  styleUrl: './login.component.css'
})
export class LoginComponent {

  username: string = '';
  password: string = '';

  constructor(
    private authService: AuthService,
    private router: Router
  ) {}

  login() {
    this.authService.authenticate(this.username, this.password).subscribe({
      next: (token) => {
        this.router.navigate(['/header'])
      },
      error: (error) => {
        console.error('Login failed:', error);
      },
    });
  }
}
